Added a scaling and default show channel and fixed drop before to be in ms

+++ charuconstruction_flutter/lib/modules/data/data.dart
@@ -109,8 +109,7 @@ class Data {
   void dropBefore(DateTime t) {
     for (final device in _devices.values) {
       device.dropBefore(
-        (t.millisecondsSinceEpoch - initialTime.millisecondsSinceEpoch)
-            .toDouble(),
+        t.millisecondsSinceEpoch - initialTime.millisecondsSinceEpoch,
       );
     }
   }
